+++ wxviews/__init__.py
@@ -2,8 +2,9 @@
 
 from pyviews.setters import import_global, inject_global, set_global, call
 from pyviews.code import Code
+from pyviews.containers import Container, View, For, If
+from pyviews.presenter import Presenter, PresenterNode, add_reference
 
-from wxviews.containers import Container, View, For, If
 from wxviews.sizers import GrowableCol, GrowableRow, set_sizer
 from wxviews.widgets import get_root, bind
 from wxviews.styles import Style, StylesView, style